+++ src/Hacks/Universal/TransitionCustomiser.cpp
@@ -23,6 +23,7 @@ class TransititionCustomiser : public Module
         {
             setID("transition-customiser");
             setCategory("Universal");
+            setPriority(6);
         }
 };