+++ src/UI/Nodes/NoRenderNode.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <Geode/Geode.hpp>
+
+namespace qolmod
+{
+    class NoRenderNode : public cocos2d::CCNode
+    {
+        public:
+            CREATE_FUNC(NoRenderNode)
+
+            virtual void visit();
+    };
+};